http4k-core/web-content/src/main/kotlin/com/example/WebContent.kt
package com.example

import org.http4k.core.Body
import org.http4k.core.ContentType.Companion.TEXT_HTML
import org.http4k.core.HttpHandler
import org.http4k.core.Method.GET
import org.http4k.core.Response
import org.http4k.core.Status.Companion.OK
import org.http4k.core.with
import org.http4k.routing.ResourceLoader
import org.http4k.routing.bind
import org.http4k.routing.routes
import org.http4k.routing.static
import org.http4k.server.SunHttp
import org.http4k.server.asServer
import org.http4k.template.HandlebarsTemplates
import org.http4k.template.ViewModel
import org.http4k.template.viewModel
import java.time.Clock
import java.time.LocalDateTime
import java.time.format.DateTimeFormatter.ISO_LOCAL_DATE_TIME

data class Home(val time: String, val browser: String) : ViewModel

fun WebContent(clock: Clock, hotReload: Boolean): HttpHandler {
    val (renderer, resourceLoader) = buildResourceLoaders(hotReload)

    return routes(
        static(resourceLoader),
        "/" bind GET to {
            val view = Body.viewModel(renderer, TEXT_HTML).toLens()

            val model = Home(LocalDateTime.now(clock).format(ISO_LOCAL_DATE_TIME), it.header("User-Agent") ?: "unknown")

            Response(OK).with(view of model)
        }
    )
}

private fun buildResourceLoaders(hotReload: Boolean) = when {
    hotReload -> HandlebarsTemplates().HotReload("./src/main/resources") to ResourceLoader.Classpath("public")
    else -> HandlebarsTemplates().CachingClasspath() to ResourceLoader.Classpath("public")
}

fun main() {
    // if setting this to true, remember to run the app with the working directory set to the root of the example
    val hotReload = false

    WebContent(Clock.systemDefaultZone(), hotReload).asServer(SunHttp(8080)).start()
}
